套餐修改接口--

+++ server/store/src/main/java/com/itheima/controller/GoodsController.java
@@ -2,6 +2,7 @@ package com.itheima.controller;
 
 
 import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.lang.UUID;
 import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
@@ -9,22 +10,26 @@ import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.itheima.dto.CommentDto;
 import com.itheima.dto.GoodsDto;
-import com.itheima.entity.Category;
-import com.itheima.entity.Comment;
-import com.itheima.entity.Goods;
+import com.itheima.entity.*;
+import com.itheima.exception.BusinessException;
 import com.itheima.service.CategoryService;
 import com.itheima.service.CommentService;
 import com.itheima.service.GoodsService;
-import com.itheima.entity.Result;
 
 import javax.annotation.Resource;
+import java.io.IOException;
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
+import com.itheima.service.SetmealService;
+import com.itheima.util.QiniuUtils;
 import com.itheima.util.Urls;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 /**
  * 商品信息(Goods)表控制层
@@ -45,6 +50,25 @@ public class GoodsController {
     private CategoryService categoryService;
     @Resource
     private CommentService commentService;
+    @Resource
+    private SetmealService setmealService;
+
+    @PostMapping(Urls.goods.upload)
+    Result Upload(@RequestBody MultipartFile imgFile) {
+        //文件全名  eg：7d104dd7-15cd-42c5-9a85-b60ea6f423c2.jpg
+        String originalFilename = imgFile.getOriginalFilename();
+        //获取后缀
+        String suffix = originalFilename.substring(originalFilename.lastIndexOf("."));
+        //随机文件名 + 后缀
+        String fileName = UUID.randomUUID().toString() + suffix;
+        //七牛云工具类 字节上传
+        try {
+            QiniuUtils.upload2Qiniu(imgFile.getBytes(), fileName);
+            return Result.success(fileName, "上传成功");
+        } catch (IOException e) {
+            throw new BusinessException("上传失败");
+        }
+    }
     @PostMapping(Urls.goods.setComment)
     public Result setComment(@RequestBody CommentDto commentDto){
         return commentService.setComment(commentDto)?Result.success("设置评论成功"):Result.fail("设置评论失败");
@@ -55,11 +79,13 @@ public class GoodsController {
         return goodsService.updateStatus(status,id)? Result.success("修改状态成功"): Result.fail("修改状态失败");
     }
     @GetMapping(Urls.goods.getPageList)
-    public Result getPageList(Long pageNum,Long pageSize,String search){
+    public Result getPageList(Long pageNum,Long pageSize,String name,String status,String categoryId){
         Page<Goods> goodsPage = new Page<>(pageNum, pageSize);
         LambdaQueryWrapper<Goods> lqw = new LambdaQueryWrapper<>();
-        lqw.like(StrUtil.isNotBlank(search),Goods::getName,search).or()
-                .like(StrUtil.isNotBlank(search),Goods::getCategoryId,search)
+        lqw.like(StrUtil.isNotBlank(name),Goods::getName,name)
+                .eq(StrUtil.isNotBlank(categoryId),Goods::getCategoryId,categoryId)
+                .eq(StrUtil.isNotBlank(status),Goods::getStatus,status)
+                .gt(Goods::getInventory,0)
                 .orderByAsc(Goods::getCreateTime).orderByDesc(Goods::getUpdateTime);
         Page<Goods> page = goodsService.page(goodsPage, lqw);
         Page<GoodsDto> goodsDtoPage = new Page<>();
@@ -70,9 +96,6 @@ public class GoodsController {
         List<Goods> records = page.getRecords();
         ArrayList<GoodsDto> goodsDtos = new ArrayList<>();
         records.stream().forEach(item->{
-            if (item.getInventory()<=0){
-                return;
-            }
             GoodsDto goodsDto = new GoodsDto();
             BeanUtil.copyProperties(item,goodsDto);
             Category byId = categoryService.getById(item.getCategoryId());
@@ -83,6 +106,7 @@ public class GoodsController {
             return Result.fail("查无此信息");
         }
         goodsDtoPage.setRecords(goodsDtos);
+        goodsDtoPage.setTotal(goodsDtos.size());
         return Result.success(goodsDtoPage,"信息查询成功");
     }
     /**
@@ -103,15 +127,20 @@ public class GoodsController {
      * @return 新增结果
      */
     @PostMapping(Urls.goods.save)
-    public Result insert(@RequestBody Goods goods) {
+    public Result insert(@RequestBody GoodsDto goods) {
         Result result = getResult(goods);
         if (result != null) {
             return result;
         }
+        String img="";
+        for (String s : goods.getImgList()) {
+            img=img+s+",";
+        }
+        goods.setImg(img);
         return goodsService.save(goods)?Result.success("添加商品成功"):Result.fail("添加商品失败");
     }
 
-    private static Result getResult(Goods goods) {
+    private static Result getResult(GoodsDto goods) {
         if (StrUtil.isBlank(goods.getName())){
             return Result.fail("商品名称不能为空");
         }
@@ -124,6 +153,9 @@ public class GoodsController {
         if (goods.getSellingPrice()<0){
             return Result.fail("商品售价输入错误");
         }
+        if (StrUtil.isBlank(goods.getImg())&&ObjectUtil.isEmpty(goods.getImgList())){
+            return Result.fail("商品展示图至少一张");
+        }
         return null;
     }
 
@@ -134,7 +166,7 @@ public class GoodsController {
      * @return 修改结果
      */
     @PutMapping(Urls.goods.update)
-    public Result update(@RequestBody Goods goods) {
+    public Result update(@RequestBody GoodsDto goods) {
         Result result = getResult(goods);
         if (result != null) {
             return result;
@@ -150,6 +182,24 @@ public class GoodsController {
      */
     @DeleteMapping(Urls.goods.delete)
     public Result delete(@RequestParam("idList") List<String> idList) {
+        // 删除图片
+        List<String> list = goodsService.listByIds(idList).stream().map(Goods::getImg).filter(StrUtil::isNotBlank).collect(Collectors.toList());
+        list.stream().forEach(item->{
+            String[] split = item.substring(0,item.lastIndexOf(",")).split(",");
+            Arrays.stream(split).forEach(QiniuUtils::deleteFileFromQiniu);
+        });
+        // 删除商品对应的评论
+        if (!goodsService.removeComment(idList)){
+            return Result.fail("删除商品评论失败");
+        }
+        // 下架商品对应的套餐
+        List<Long> setmealId = goodsService.getSetmealId(idList);
+        if (!setmealId.isEmpty()){
+            List<Long> collect = setmealId.stream().distinct().collect(Collectors.toList());
+            if (!setmealService.updateStatus(collect)){
+                return Result.fail("下架套餐失败");
+            }
+        }
         return goodsService.removeByIds(idList)?Result.success("删除商品成功"):Result.fail("删除商品失败");
     }
 }
